Update of registration Callback using the .register(this, SENDER_ID); method
In the registration device section, initially is reported the registration using the method with context and SENDER_ID parameters.

Push.with(session).register(this, SENDER_ID);

Right after (where async callback is explained) is used the other one version where the register is invoked (passing the registrationId parameter). It's not immediate understand that registrationId parameter is returned from a previous (or manual) registeration.